test(e2e): stabilize steer screenshot capture
